Move bot endpoints to `/bot` path
