+++ packages/api/src/routes/bot/utils.ts
