+++ packages/api/test/routes/bot/code/download.test.ts
@@ -9,10 +9,10 @@ import * as tar from 'tar';
 
 import { bots, orgs, repos, tasks } from '@automa/prisma';
 
-import { getStubs } from '../../mocks';
-import { call, seedBots, seedOrgs, seedRepos, server } from '../../utils';
+import { getStubs } from '../../../mocks';
+import { call, seedBots, seedOrgs, seedRepos, server } from '../../../utils';
 
-suite('code/download', () => {
+suite('bot/code/download', () => {
   let app: FastifyInstance, response: LightMyRequestResponse;
   let org: orgs, bot: bots, repo: repos, task: tasks;
   let sandbox: SinonSandbox, postStub: SinonStub, tarCreateStub: SinonStub;
@@ -820,7 +820,7 @@ const download = async (
   app: FastifyInstance,
   task: Pick<tasks, 'id' | 'token'>,
 ) =>
-  call(app, '/code/download', {
+  call(app, '/bot/code/download', {
     method: 'POST',
     payload: {
       task,
