+++ packages/api/test/routes/bot/code/propose.test.ts
@@ -7,8 +7,8 @@ import axios from 'axios';
 
 import { bots, orgs, repos, tasks } from '@automa/prisma';
 
-import { getStubs } from '../../mocks';
-import { call, seedBots, seedOrgs, seedRepos, server } from '../../utils';
+import { getStubs } from '../../../mocks';
+import { call, seedBots, seedOrgs, seedRepos, server } from '../../../utils';
 
 const diff = [
   'diff --git a/file.txt b/file.txt',
@@ -20,7 +20,7 @@ const diff = [
   '+new',
 ].join('\n');
 
-suite('code/propose', () => {
+suite('bot/code/propose', () => {
   let app: FastifyInstance, response: LightMyRequestResponse;
   let org: orgs, bot: bots, repo: repos, task: tasks;
   let sandbox: SinonSandbox, postStub: SinonStub, getStub: SinonStub;
@@ -116,7 +116,7 @@ suite('code/propose', () => {
 
   suite('with empty body', () => {
     setup(async () => {
-      response = await call(app, '/code/propose', {
+      response = await call(app, '/bot/code/propose', {
         method: 'POST',
         payload: {},
       });
@@ -321,7 +321,7 @@ suite('code/propose', () => {
 
   suite('with missing proposal token', () => {
     setup(async () => {
-      response = await call(app, '/code/propose', {
+      response = await call(app, '/bot/code/propose', {
         method: 'POST',
         payload: {
           task: {
@@ -367,7 +367,7 @@ suite('code/propose', () => {
 
   suite('with null proposal token', () => {
     setup(async () => {
-      response = await call(app, '/code/propose', {
+      response = await call(app, '/bot/code/propose', {
         method: 'POST',
         payload: {
           task: {
@@ -414,7 +414,7 @@ suite('code/propose', () => {
 
   suite('with missing proposal diff', () => {
     setup(async () => {
-      response = await call(app, '/code/propose', {
+      response = await call(app, '/bot/code/propose', {
         method: 'POST',
         payload: {
           task: {
@@ -460,7 +460,7 @@ suite('code/propose', () => {
 
   suite('with null proposal diff', () => {
     setup(async () => {
-      response = await call(app, '/code/propose', {
+      response = await call(app, '/bot/code/propose', {
         method: 'POST',
         payload: {
           task: {
@@ -2536,7 +2536,7 @@ const propose = async (
   },
   metadata?: Record<string, any>,
 ) =>
-  call(app, '/code/propose', {
+  call(app, '/bot/code/propose', {
     method: 'POST',
     payload: {
       task,
